+++ refactor/ast.py
@@ -5,7 +5,9 @@ from collections import UserList
 from contextlib import contextmanager
 from dataclasses import dataclass
 from functools import cached_property
-from typing import Any, List, Protocol
+from typing import Any, Generator, List, Optional, Protocol, Tuple, Union
+
+from refactor import common
 
 
 @dataclass
@@ -54,7 +56,7 @@ class Unparser(Protocol):
         ...  # pragma: no cover
 
 
-class UnparserBase(ast._Unparser):  # type: ignore
+class BaseUnparser(ast._Unparser):  # type: ignore
     # Normally ast._Unparser is a private API
     # though since it doesn't tend to change
     # often, we could simply have a base class
@@ -62,25 +64,89 @@ class UnparserBase(ast._Unparser):  # type: ignore
     # incompatible changes and let the refactor
     # users to override it.
 
-    def __init__(self, source: str, *args: Any, **kwargs: Any) -> None:
-        self.source = source
+    source: Optional[str]
+
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        self.source = kwargs.pop("source", None)
         super().__init__(*args, **kwargs)
 
     def unparse(self, node: ast.AST) -> str:
         return self.visit(node)
 
     @cached_property
-    def tokens(self):
+    def tokens(self) -> Tuple[tokenize.TokenInfo, ...]:
         buffer = io.StringIO(self.source)
         token_stream = tokenize.generate_tokens(buffer.readline)
         return tuple(token_stream)
 
-    @cached_property
-    def token_map(self):
-        return {(*token.start, *token.end): token for token in self.tokens}
-
     @contextmanager
-    def indented(self):
+    def indented(self) -> Generator[None, None, None]:
         self._indent += 1
         yield
         self._indent -= 1
+
+
+class PreciseUnparser(BaseUnparser):
+    """
+    Try to locate precise textual versions of child nodes by
+    bi-directional AST equivalence with the versions that exist
+    on the source.
+    """
+
+    def traverse(self, node: Union[List[ast.AST], ast.AST]) -> None:
+        if isinstance(node, list) or self.source is None:
+            return super().traverse(node)
+
+        did_retrieve = self.maybe_retrieve(node)
+        if not did_retrieve:
+            super().traverse(node)
+
+    def maybe_retrieve(self, node: ast.AST) -> bool:
+        # Process:
+        #   - Check whether the unparser has access to the
+        #     current source code.
+        #   - Check whether the given node is an expression
+        #     or statement.
+        #   - Try retrieving the source segment
+        #   - If that succeeds, try parsing that segment
+        #   - Ensure the ASTs are identical
+        #   - Write-off the original source
+
+        if self.source is None:
+            return False
+
+        if not isinstance(node, (ast.stmt, ast.expr)):
+            return False
+
+        segment = common.get_source_segment(self.source, node)
+        if segment is None:
+            return False
+
+        try:
+            tree = ast.parse(segment)
+        except SyntaxError:
+            return False
+
+        if len(tree.body) != 1:
+            return False
+
+        [retrieved_node] = tree.body
+
+        # If this is a pure expression, then unpack
+        # the actual value.
+        if isinstance(node, ast.expr):
+            retrieved_node = retrieved_node.value
+
+        if is_same_ast := common.compare_ast(retrieved_node, node):
+            self.retrieve_segment(node, segment)
+
+        return is_same_ast
+
+    def retrieve_segment(self, node: ast.AST, segment: str) -> None:
+        if isinstance(node, ast.stmt):
+            self.fill()
+
+        self.write(segment)
+
+
+UNPARSER_BACKENDS = {"fast": BaseUnparser, "precise": PreciseUnparser}